fixing check for variable existence
By trying to call enc1 and enc2 with them possibly not being defined it
can throw a NameError. In addition, when it is defined, this will always
evaluate to true unless the /value/ of enc1 and enc2 are strings that
are the names of variables. locals() returns a dictionary, where the
keys are the names of the variable as type string and the values are the
values of the variables. The search must be performed with a string
instead of by type. This also avoids the earlier mentioned error.